Add support to multi-row.
Multi-row allows to inherit fields from a previous HTML element (<tr>)
which contains information about subsequent elements (<td>). This is
common in trackers that display items in a "grouped" layout.

This feature extends (and replace the "dateheaders" feature) allowing to
inherit more fields.

Compare the tests TestIndexerDefinitionRunner_MultiRowSearch and
TestIndexerDefinitionRunner_DateHeadersSearch to see how to replace the
"dateheaders".